完成了评论回复国内

+++ src/main/java/com/littlekingkong/community/controller/PublishController.java
@@ -36,7 +36,7 @@ public class PublishController {
     public String publish(@RequestParam(value = "title", required = false) String title,
                           @RequestParam(value = "description", required = false) String description,
                           @RequestParam(value = "tag", required = false) String tag,
-                          @RequestParam(value = "id", required = false) Integer id,
+                          @RequestParam(value = "id", required = false) Long id,
                           @RequestParam(value = "comment_count",defaultValue = "0",required = false) Integer comment_count,
                           @RequestParam(value = "view_count",defaultValue = "0",required = false) Integer view_count,
                           @RequestParam(value = "like_count",defaultValue = "0",required = false) Integer like_count,
@@ -83,7 +83,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
 
         Question question = questionService.getQuestionById(id);
